requirements.txt
# Core dependencies
openai==1.3.5
openai-whisper==20231117
python-dotenv==1.0.0

# Zoom integration
requests==2.31.0
websockets==12.0
aiohttp==3.9.0
PyJWT==2.8.0

# Audio processing
pydub==0.25.1
torch==2.1.1
torchaudio==2.1.1

# Optional: For enhanced features
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23

# Development
pytest==7.4.3
black==23.11.0